fix(apptainer): ensure apptainer jobs run as privileged
Apptainer jobs require privileged access to function correctly, for
instance to mount volumes and manage cgroups.

The previous implementation incorrectly relied on `job.is_containerized`
to enable privileged mode for Apptainer jobs. I suspect this flag is
intended for use with the `--containerize` option and does not reliably
indicate that a job is running in a container.

This commit enables privileged mode whenever the Apptainer deployment
method is used.